refactor: extract Station entity to domain layer
- Move Station struct and implementation to src/domain/entities/station.rs
- Add re-export in entities.rs for backward compatibility
- Follow clean architecture principles with domain entity isolation
