+++ src/domain/entities/mod.rs
@@ -1,1 +1,1 @@
-// TODO: Extract entities here
+pub mod station;
